refactor users
